Enum macro_utils::field::FieldName

pub enum FieldName {
    Ident(Ident),
    Index(Index),
}

Represent the way to access a field. Either with Self::Ident or Self::Index

Variants

Ident(Ident)

the field is accessed with an ident as name field

Index(Index)

the field is accessed with an index as a tuple or uname field

Implementations

impl FieldName

pub fn from_field_ref(field: &Field) -> Self

Get the way to access the field from a Field. It clones only field.ident and therefore less resource intensive than cloning the whole field in Self::from_field using Self::from_field(field.clone()).

pub fn from_field(field: Field) -> Self

Get the way to access the field from a Field. If you want to reuse the field afterward use Self::from_field_ref instead.

pub const fn require_ident(&self) -> Option<&Ident>

Return Some on an Ident and None on an Index